requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.5
cfgv==3.2.0
distlib==0.3.1
Django==3.1.6
django-dotenv==1.4.2
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
identify==1.6.1
importlib-metadata==3.7.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
nodeenv==1.5.0
pre-commit==2.10.1
pycodestyle==2.6.0
pylint==2.6.0
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
virtualenv==20.4.2
wrapt==1.12.1
zipp==3.4.0
